\name{mumocomp-package}
\alias{mumocomp-package}
\alias{mumocomp}
\docType{package}
\title{
\packageTitle{mumocomp}
}
\description{
\packageDescription{mumocomp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mumocomp}
\packageIndices{mumocomp}
}
\author{
\packageAuthor{mumocomp}

Maintainer: \packageMaintainer{mumocomp}
}
\references{
}
\keyword{ htest }
\seealso{
}
\examples{
}
