\name{mumotest}
\alias{mumotest}
\title{Simultaneous Inference for parameters in Multiple Models}
\description{Blah, blah, blah}
\usage{
mumotest(mlist, K, B)
}
\arguments{
  \item{mlist}{List of model objects}
  \item{K}{A contrast matrix}
  \item{B}{Number of iterations}
}

\value{A vector with p-values}
\author{Daniel Gerhard}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ htest }% use one of  RShowDoc("KEYWORDS")
